requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements.txt requirements.in
#


cffi==1.9.1               # via cryptography
click==6.7                # via flask
cryptography==1.7.1       # via pyopenssl
enum34==1.1.6             # via cryptography
flask==0.12
idna==2.2                 # via cryptography
ipaddress==1.0.17         # via cryptography
itsdangerous==0.24        # via flask
Jinja2==2.9.3             # via flask
MarkupSafe==0.23          # via jinja2
pyasn1==0.1.9             # via cryptography
pycparser==2.17           # via cffi
PyJWT==1.4.2
pyOpenSSL==16.2.0
requests==2.12.4
six==1.10.0               # via cryptography, pyopenssl
Werkzeug==0.11.15         # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via cryptography
